Updated performance tests

+++ example/DemoSources/PixelTest.hpp
@@ -21,7 +21,7 @@ namespace PixelTest
 		lwmf::FilledRectangle(ScreenTexture, 0, 65, ScreenTexture.Width, 20, 0, 0);
 
 		std::array<char, 15> MPixelsPerSecond{};
-		std::to_chars(MPixelsPerSecond.data(), MPixelsPerSecond.data() + MPixelsPerSecond.size(), static_cast<float>(ScreenTexture.Size * lwmf::FPS) / 1000000.0F );
+		std::to_chars(MPixelsPerSecond.data(), MPixelsPerSecond.data() + MPixelsPerSecond.size(), static_cast<float>(static_cast<std::uint_fast64_t>(ScreenTexture.Size) * static_cast<std::uint_fast64_t>(lwmf::FPS)) / 1000000.0F );
 		lwmf::RenderText(ScreenTexture, "Throughput in MegaPixels/s: " + std::string(MPixelsPerSecond.data()), 10, 70, 0xFFFFFFFF);
 	}
